Adding Welcome Page, Token Input Page
