+++ components/interview/NoAccessPage.tsx
@@ -1,0 +1,7 @@
+export function NoAccessPage() {
+    return (
+        <div>
+            <h1>Sorry, either your session hasn't started yet or you do not have access to this session. Please check your token and try again.</h1>
+        </div>
+    );
+}
